dump_unused_files/build_artifacts/updated_packages/requirements.txt
# CodeGrey AI SOC Platform - Dynamic Edition
# NO HARDCODING - ALL DYNAMIC FEATURES

# Core dependencies
requests>=2.31.0
pyyaml>=6.0
psutil>=5.9.0
aiohttp>=3.8.0

# Network discovery
python-nmap>=0.7.1
scapy>=2.5.0

# Machine learning for dynamic detection
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Container orchestration
docker>=6.1.0

# Dynamic threat analysis
nltk>=3.8.0

# Windows-specific (for Windows package)
pywin32>=306; sys_platform == "win32"
wmi>=1.5.1; sys_platform == "win32"

# Linux-specific (for Linux package)
pyinotify>=0.9.6; sys_platform == "linux"

# Logging and monitoring
loguru>=0.7.0

# Security
cryptography>=41.0.0

# Async support
aiofiles>=23.0.0
